Added pre and post processors to resource handling.
app|manager.preprocessor and app|manager.postprocessor point to a
pre-processor and post-processor respectively for each managed resource.
Both objects may be used as decorators to collect pre/post processor
functions (which are called in the order they are collected). Either may
be replaced by any object with a .process() method that takes the
request data (pre-processor) or response data (post-processor) as the
first argument, and a resource as the second. Pre-processors are able to
trigger an immediate response, by returning any value other than None
(which will then be used as the response data). Post-processors must
always return a value, which will be used as the response data.