Added some utilities
Added Watcher and Cache to findig.utils

Watcher watches a manager (or app) for 'hit', 'read', 'delete', 'create'
and 'modified' events on any resource.

Cache intercepts GET requests and avoids calling the resource getter
function whenever the valid data is cached for the resource. Cache uses
an in-memory dictionary and is not intended for production use; other
back-ends can be supported subclassing Cache and implementing
__contains__, __getitem__, __setitem__ and __delitem__.