Added parameter method_hack to Resource
When enabled, a resource will listen for GET and POST requests on
/url/put, /url/patch and /url/delete respectively and internally
restructure those requests to look like PUT, PATCH and DELETE requests
respectively that were sent to /url.

It is mainly useful as an easy workaround to allow web-browsers that
don't send PUT PATCH and DELETE requests from web forms to fully access
the resource.