Direct template rendering for TemplateFormatter
Added render_template to TemplateFormatter, which uses a couple hacks to
render a template directly by name.

Note: It passes the data through the entire formatting chain, which
means that in some cases the template may not be rendered, but another
response type might be used depending on the formatter's configuration
(example: if the formatter has a separate formatter function registered
for "application/json" responses, and the Accept header prefers
appplication/json, then the data will be sent straight to the formatter
function, ignoring the template).