Made CollectionResource aware of what it collects
Improvement to CollectionResource optionally makes it aware of the
Resource that it collects. As a result, CollectionResource can now
automatically handle POST and DELETE requests for the collection, using
the saver and deleter functions from the collected resource.

Added the resource.collection() decorator to mark a CollectionResource
as collecting resource.