Added better context management for Apps
The decorator @app.context allows the definition of a context manager
than can setup and teardown application resources. The function must
yield after setup code and before teardown code. If it yields a value,
then that value will be set on findig.context.ctx with the function's
name. For example this code will set ctx.db for the duration of the
application context::

@app.context
def db():
# Db setup code
yield db
# Db teardown code

Additionally, a second decorator @app.cleanup_hook allows you to
register a regular function that gets called at the end of the
application context to cleanup resources. This is useful when a resource
is created after
the application context is established but still needs to be torn down
before the app exits.