Fixed bug in the Counter tool
This particular error affects the default Counter storage class, which
was rather careless about how it stored its hits. It attempted to store
them in a sorted heap as tuples, but parts of that tuple could vary
types depending on the requests, and cause an error to the raised when
trying to compare the differing types.

The fix for this issue now restricts sorting the hits /only/ by the time
they are received, which is meant to be always comparable.